[balancer] Use existing nemo-keepalive, port to C++
We already have a library in Nemo Mobile that takes care of keepalive. No need
to write a separate tool for it. Also, the shell script had some problems that
might have caused the keepalive tool to run forever, so rewrite the tool in
C++, plus running "btrfs" twice just to grep values and then put those into
awk and calculate the usage percentage is.. hmm.